[refactor] user에서 follow 서비스 패키지 분리 (#80)
