+++ src/main/java/konkuk/thip/user/application/service/follow/UserGetFollowersService.java
@@ -1,4 +1,4 @@
-package konkuk.thip.user.application.service;
+package konkuk.thip.user.application.service.follow;
 
 import konkuk.thip.user.adapter.in.web.response.UserFollowersResponse;
 import konkuk.thip.user.application.port.in.UserGetFollowersUsecase;
